+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/FilteredTable.java
@@ -20,6 +20,7 @@ import is.codion.swing.common.model.component.table.FilteredTableSearchModel.Row
 import is.codion.swing.common.model.component.table.FilteredTableSelectionModel;
 import is.codion.swing.common.model.component.table.FilteredTableSortModel;
 import is.codion.swing.common.ui.KeyEvents;
+import is.codion.swing.common.ui.KeyboardShortcut;
 import is.codion.swing.common.ui.Utilities;
 import is.codion.swing.common.ui.border.Borders;
 import is.codion.swing.common.ui.component.Components;
@@ -36,6 +37,7 @@ import javax.swing.JComponent;
 import javax.swing.JTable;
 import javax.swing.JTextField;
 import javax.swing.JViewport;
+import javax.swing.KeyStroke;
 import javax.swing.ListSelectionModel;
 import javax.swing.SortOrder;
 import javax.swing.table.JTableHeader;
@@ -55,7 +57,9 @@ import java.text.Collator;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Comparator;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.ResourceBundle;
 import java.util.function.Consumer;
@@ -65,9 +69,11 @@ import java.util.stream.Stream;
 import static is.codion.common.item.Item.item;
 import static is.codion.swing.common.model.component.combobox.ItemComboBoxModel.itemComboBoxModel;
 import static is.codion.swing.common.model.component.table.FilteredTableSortModel.nextSortOrder;
+import static is.codion.swing.common.ui.KeyboardShortcut.keyStrokeValue;
 import static is.codion.swing.common.ui.component.Components.borderLayoutPanel;
 import static is.codion.swing.common.ui.component.Components.itemComboBox;
 import static is.codion.swing.common.ui.component.table.ColumnConditionPanel.columnConditionPanel;
+import static is.codion.swing.common.ui.component.table.FilteredTable.KeyboardShortcuts.*;
 import static is.codion.swing.common.ui.component.table.FilteredTableConditionPanel.filteredTableConditionPanel;
 import static is.codion.swing.common.ui.control.Control.control;
 import static java.awt.event.InputEvent.*;
@@ -102,6 +108,22 @@ public final class FilteredTable<R, C> extends JTable {
           Configuration.booleanValue("is.codion.swing.common.ui.component.table.FilteredTable.allowColumnReordering", true);
 
   private static final ResourceBundle MESSAGES = ResourceBundle.getBundle(FilteredTable.class.getName());
+
+  /**
+   * The available keyboard shortcuts.
+   */
+  public enum KeyboardShortcuts implements KeyboardShortcut {
+    COPY_CELL,
+    TOGGLE_SORT_COLUMN,
+    TOGGLE_SORT_COLUMN_ADD;
+
+    private static final Map<KeyboardShortcut, Value<KeyStroke>> KEYSTROKES = createDefaultKeystrokes();
+
+    @Override
+    public Value<KeyStroke> keyStroke() {
+      return KEYSTROKES.get(this);
+    }
+  }
 
   /**
    * Specifies whether to center the scrolled to row and or column.
@@ -657,16 +679,13 @@ public final class FilteredTable<R, C> extends JTable {
     tableModel.sortModel().addSortingChangedListener(columnIdentifier -> getTableHeader().repaint());
     addMouseListener(new FilteredTableMouseListener());
     addKeyListener(new MoveResizeColumnKeyListener());
-    KeyEvents.builder(VK_C)
-            .modifiers(CTRL_DOWN_MASK | ALT_DOWN_MASK)
+    KeyEvents.builder(COPY_CELL.keyStroke().get())
             .action(control(this::copySelectedCell))
             .enable(this);
-    KeyEvents.builder(VK_UP)
-            .modifiers(ALT_DOWN_MASK)
+    KeyEvents.builder(TOGGLE_SORT_COLUMN_ADD.keyStroke().get())
             .action(control(() -> toggleColumnSorting(getSelectedColumn(), true)))
             .enable(this);
-    KeyEvents.builder(VK_DOWN)
-            .modifiers(ALT_DOWN_MASK)
+    KeyEvents.builder(TOGGLE_SORT_COLUMN.keyStroke().get())
             .action(control(() -> toggleColumnSorting(getSelectedColumn(), false)))
             .enable(this);
   }
@@ -684,6 +703,15 @@ public final class FilteredTable<R, C> extends JTable {
     if (object instanceof JComponent) {
       components.add((JComponent) object);
     }
+  }
+
+  private static Map<KeyboardShortcut, Value<KeyStroke>> createDefaultKeystrokes() {
+    Map<KeyboardShortcut, Value<KeyStroke>> keyStrokes = new HashMap<>();
+    keyStrokes.put(COPY_CELL, keyStrokeValue(VK_C, CTRL_DOWN_MASK | ALT_DOWN_MASK));
+    keyStrokes.put(TOGGLE_SORT_COLUMN, keyStrokeValue(VK_DOWN, ALT_DOWN_MASK));
+    keyStrokes.put(TOGGLE_SORT_COLUMN_ADD, keyStrokeValue(VK_UP, ALT_DOWN_MASK));
+
+    return keyStrokes;
   }
 
   /**