+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityPanel.java
@@ -11,6 +11,7 @@ import is.codion.common.value.Value;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.domain.entity.EntityType;
 import is.codion.swing.common.ui.KeyEvents;
+import is.codion.swing.common.ui.KeyboardShortcut;
 import is.codion.swing.common.ui.Utilities;
 import is.codion.swing.common.ui.Windows;
 import is.codion.swing.common.ui.component.Components;
@@ -29,6 +30,7 @@ import javax.swing.AbstractAction;
 import javax.swing.JComponent;
 import javax.swing.JPanel;
 import javax.swing.JToolBar;
+import javax.swing.KeyStroke;
 import javax.swing.SwingConstants;
 import javax.swing.WindowConstants;
 import java.awt.BorderLayout;
@@ -42,17 +44,21 @@ import java.beans.PropertyChangeEvent;
 import java.beans.PropertyChangeListener;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.ResourceBundle;
 import java.util.function.Consumer;
 
+import static is.codion.swing.common.ui.KeyboardShortcut.keyStrokeValue;
 import static is.codion.swing.common.ui.Utilities.parentOfType;
 import static is.codion.swing.common.ui.Utilities.parentWindow;
 import static is.codion.swing.common.ui.component.Components.*;
 import static is.codion.swing.common.ui.component.button.ToggleButtonType.CHECKBOX;
 import static is.codion.swing.common.ui.layout.Layouts.borderLayout;
 import static is.codion.swing.framework.ui.EntityPanel.Direction.*;
+import static is.codion.swing.framework.ui.EntityPanel.KeyboardShortcuts.*;
 import static is.codion.swing.framework.ui.EntityPanel.PanelState.*;
 import static java.awt.event.InputEvent.*;
 import static java.awt.event.KeyEvent.*;
@@ -167,6 +173,34 @@ public class EntityPanel extends JPanel {
    */
   public enum Direction {
     UP, DOWN, RIGHT, LEFT
+  }
+
+  /**
+   * The keyboard shortcuts available for {@link EntityPanel}s.
+   * Note that changing the shortcut keystroke after the panel
+   * has been initialized has no effect.
+   */
+  public enum KeyboardShortcuts implements KeyboardShortcut {
+    REQUEST_TABLE_FOCUS,
+    TOGGLE_CONDITION_PANEL,
+    SELECT_CONDITION_PANEL,
+    TOGGLE_FILTER_PANEL,
+    SELECT_FILTER_PANEL,
+    REQUEST_SEARCH_FIELD_FOCUS,
+    REQUEST_EDIT_PANEL_FOCUS,
+    SELECT_INPUT_FIELD,
+    TOGGLE_EDIT_PANEL,
+    NAVIGATE_UP,
+    NAVIGATE_DOWN,
+    NAVIGATE_RIGHT,
+    NAVIGATE_LEFT;
+
+    private static final Map<KeyboardShortcut, Value<KeyStroke>> KEYSTROKES = createDefaultKeystrokes();
+
+    @Override
+    public Value<KeyStroke> keyStroke() {
+      return KEYSTROKES.get(this);
+    }
   }
 
   private final SwingEntityModel entityModel;
@@ -737,105 +771,103 @@ public class EntityPanel extends JPanel {
   protected final void setupKeyboardActions() {
     if (containsTablePanel()) {
       tablePanel.control(TableControl.REQUEST_TABLE_FOCUS).optional().ifPresent(control ->
-              KeyEvents.builder(VK_T)
-                      .modifiers(CTRL_DOWN_MASK)
+              KeyEvents.builder(REQUEST_TABLE_FOCUS.keyStroke().get())
                       .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
                       .action(control)
                       .enable(this));
       tablePanel.control(TableControl.TOGGLE_CONDITION_PANEL).optional().ifPresent(control ->
-              KeyEvents.builder(VK_S)
-                      .modifiers(CTRL_DOWN_MASK | ALT_DOWN_MASK)
+              KeyEvents.builder(TOGGLE_CONDITION_PANEL.keyStroke().get())
                       .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
                       .action(control)
                       .enable(this));
       tablePanel.control(TableControl.SELECT_CONDITION_PANEL).optional().ifPresent(control ->
-              KeyEvents.builder(VK_S)
-                      .modifiers(CTRL_DOWN_MASK)
+              KeyEvents.builder(SELECT_CONDITION_PANEL.keyStroke().get())
                       .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
                       .action(control)
                       .enable(this));
       tablePanel.control(TableControl.TOGGLE_FILTER_PANEL).optional().ifPresent(control ->
-              KeyEvents.builder(VK_F)
-                      .modifiers(CTRL_DOWN_MASK | ALT_DOWN_MASK)
+              KeyEvents.builder(TOGGLE_FILTER_PANEL.keyStroke().get())
                       .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
                       .action(control)
                       .enable(this));
       tablePanel.control(TableControl.SELECT_FILTER_PANEL).optional().ifPresent(control ->
-              KeyEvents.builder(VK_F)
-                      .modifiers(CTRL_DOWN_MASK | SHIFT_DOWN_MASK)
+              KeyEvents.builder(SELECT_FILTER_PANEL.keyStroke().get())
                       .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
                       .action(control)
                       .enable(this));
-      KeyEvents.builder(VK_F)
-              .modifiers(CTRL_DOWN_MASK)
+      KeyEvents.builder(REQUEST_SEARCH_FIELD_FOCUS.keyStroke().get())
               .action(createRequestTableSearchFieldControl())
               .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
               .enable(this);
       if (containsEditPanel()) {
         tablePanel.control(TableControl.REQUEST_TABLE_FOCUS).optional().ifPresent(control ->
-                KeyEvents.builder(VK_T)
-                        .modifiers(CTRL_DOWN_MASK)
+                KeyEvents.builder(REQUEST_TABLE_FOCUS.keyStroke().get())
                         .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
                         .action(control)
                         .enable(editControlPanel));
         tablePanel.control(TableControl.TOGGLE_CONDITION_PANEL).optional().ifPresent(control ->
-                KeyEvents.builder(VK_S)
-                        .modifiers(CTRL_DOWN_MASK | ALT_DOWN_MASK)
+                KeyEvents.builder(TOGGLE_CONDITION_PANEL.keyStroke().get())
                         .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
                         .action(control)
                         .enable(editControlPanel));
         tablePanel.control(TableControl.SELECT_CONDITION_PANEL).optional().ifPresent(control ->
-                KeyEvents.builder(VK_S)
-                        .modifiers(CTRL_DOWN_MASK)
+                KeyEvents.builder(SELECT_CONDITION_PANEL.keyStroke().get())
                         .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
                         .action(control)
                         .enable(editControlPanel));
         tablePanel.control(TableControl.TOGGLE_FILTER_PANEL).optional().ifPresent(control ->
-                KeyEvents.builder(VK_F)
-                        .modifiers(CTRL_DOWN_MASK | ALT_DOWN_MASK)
+                KeyEvents.builder(TOGGLE_FILTER_PANEL.keyStroke().get())
                         .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
                         .action(control)
                         .enable(editControlPanel));
         tablePanel.control(TableControl.SELECT_FILTER_PANEL).optional().ifPresent(control ->
-                KeyEvents.builder(VK_F)
-                        .modifiers(CTRL_DOWN_MASK | SHIFT_DOWN_MASK)
+                KeyEvents.builder(SELECT_FILTER_PANEL.keyStroke().get())
                         .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
                         .action(control)
                         .enable(editControlPanel));
       }
     }
     if (containsEditPanel()) {
-      Control requestEditPanelFocus = createRequestEditPanelFocusControl();
-      Control selectInputComponent = createSelectInputComponentControl();
-      KeyEvents.builder(VK_E)
-              .modifiers(CTRL_DOWN_MASK)
+      KeyEvents.builder(REQUEST_EDIT_PANEL_FOCUS.keyStroke().get())
               .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
-              .action(requestEditPanelFocus)
+              .action(createRequestEditPanelFocusControl())
               .enable(this);
-      KeyEvents.builder(VK_I)
-              .modifiers(CTRL_DOWN_MASK)
+      KeyEvents.builder(SELECT_INPUT_FIELD.keyStroke().get())
               .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
-              .action(selectInputComponent)
-              .enable(this);
-      KeyEvents.builder(VK_I)
-              .modifiers(CTRL_DOWN_MASK)
+              .action(createSelectInputComponentControl())
+              .enable(this, editControlPanel);
+      KeyEvents.builder(TOGGLE_EDIT_PANEL.keyStroke().get())
               .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
-              .action(selectInputComponent)
-              .enable(editControlPanel);
-      Control toggleEditPanel = createToggleEditPanelControl();
-      KeyEvents.builder(VK_E)
-              .modifiers(CTRL_DOWN_MASK | ALT_DOWN_MASK)
-              .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
-              .action(toggleEditPanel)
-              .enable(this);
-      KeyEvents.builder(VK_E)
-              .modifiers(CTRL_DOWN_MASK | ALT_DOWN_MASK)
-              .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
-              .action(toggleEditPanel)
-              .enable(editControlPanel);
+              .action(createToggleEditPanelControl())
+              .enable(this, editControlPanel);
     }
     if (settings.useKeyboardNavigation) {
       setupNavigation();
+    }
+  }
+
+  protected final void setupNavigation() {
+    KeyEvents.Builder navigateUp = KeyEvents.builder(NAVIGATE_UP.keyStroke().get())
+            .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
+            .action(new Navigate(UP));
+    KeyEvents.Builder navigateDown = KeyEvents.builder(NAVIGATE_DOWN.keyStroke().get())
+            .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
+            .action(new Navigate(DOWN));
+    KeyEvents.Builder navigateRight = KeyEvents.builder(NAVIGATE_RIGHT.keyStroke().get())
+            .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
+            .action(new Navigate(RIGHT));
+    KeyEvents.Builder navigateLeft = KeyEvents.builder(NAVIGATE_LEFT.keyStroke().get())
+            .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
+            .action(new Navigate(LEFT));
+    navigateUp.enable(this);
+    navigateDown.enable(this);
+    navigateRight.enable(this);
+    navigateLeft.enable(this);
+    if (containsEditPanel()) {
+      navigateUp.enable(editControlPanel);
+      navigateDown.enable(editControlPanel);
+      navigateRight.enable(editControlPanel);
+      navigateLeft.enable(editControlPanel);
     }
   }
 
@@ -881,51 +913,6 @@ public class EntityPanel extends JPanel {
             .mnemonic(Messages.refreshMnemonic())
             .smallIcon(FrameworkIcons.instance().refresh())
             .build();
-  }
-
-  protected final void setupNavigation() {
-    KeyEvents.builder(VK_UP)
-            .modifiers(ALT_DOWN_MASK | CTRL_DOWN_MASK)
-            .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
-            .action(new Navigate(UP))
-            .enable(this);
-    KeyEvents.builder(VK_DOWN)
-            .modifiers(ALT_DOWN_MASK | CTRL_DOWN_MASK)
-            .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
-            .action(new Navigate(DOWN))
-            .enable(this);
-    KeyEvents.builder(VK_RIGHT)
-            .modifiers(ALT_DOWN_MASK | CTRL_DOWN_MASK)
-            .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
-            .action(new Navigate(RIGHT))
-            .enable(this);
-    KeyEvents.builder(VK_LEFT)
-            .modifiers(ALT_DOWN_MASK | CTRL_DOWN_MASK)
-            .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
-            .action(new Navigate(LEFT))
-            .enable(this);
-    if (containsEditPanel()) {
-      KeyEvents.builder(VK_UP)
-              .modifiers(ALT_DOWN_MASK | CTRL_DOWN_MASK)
-              .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
-              .action(new Navigate(UP))
-              .enable(editControlPanel);
-      KeyEvents.builder(VK_DOWN)
-              .modifiers(ALT_DOWN_MASK | CTRL_DOWN_MASK)
-              .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
-              .action(new Navigate(DOWN))
-              .enable(editControlPanel);
-      KeyEvents.builder(VK_RIGHT)
-              .modifiers(ALT_DOWN_MASK | CTRL_DOWN_MASK)
-              .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
-              .action(new Navigate(RIGHT))
-              .enable(editControlPanel);
-      KeyEvents.builder(VK_LEFT)
-              .modifiers(ALT_DOWN_MASK | CTRL_DOWN_MASK)
-              .condition(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
-              .action(new Navigate(LEFT))
-              .enable(editControlPanel);
-    }
   }
 
   protected final void initializeEditPanel() {
@@ -1099,6 +1086,25 @@ public class EntityPanel extends JPanel {
     if (initialized) {
       throw new IllegalStateException("Method must be called before the panel is initialized");
     }
+  }
+
+  private static Map<KeyboardShortcut, Value<KeyStroke>> createDefaultKeystrokes() {
+    Map<KeyboardShortcut, Value<KeyStroke>> keyStrokes = new HashMap<>();
+    keyStrokes.put(REQUEST_TABLE_FOCUS, keyStrokeValue(VK_T, CTRL_DOWN_MASK));
+    keyStrokes.put(TOGGLE_CONDITION_PANEL, keyStrokeValue(VK_S, CTRL_DOWN_MASK | ALT_DOWN_MASK));
+    keyStrokes.put(SELECT_CONDITION_PANEL, keyStrokeValue(VK_S, CTRL_DOWN_MASK));
+    keyStrokes.put(TOGGLE_FILTER_PANEL, keyStrokeValue(VK_F, CTRL_DOWN_MASK | ALT_DOWN_MASK));
+    keyStrokes.put(SELECT_FILTER_PANEL, keyStrokeValue(VK_F, CTRL_DOWN_MASK | SHIFT_DOWN_MASK));
+    keyStrokes.put(REQUEST_SEARCH_FIELD_FOCUS, keyStrokeValue(VK_F, CTRL_DOWN_MASK));
+    keyStrokes.put(REQUEST_EDIT_PANEL_FOCUS, keyStrokeValue(VK_E, CTRL_DOWN_MASK));
+    keyStrokes.put(SELECT_INPUT_FIELD, keyStrokeValue(VK_I, CTRL_DOWN_MASK));
+    keyStrokes.put(TOGGLE_EDIT_PANEL, keyStrokeValue(VK_E, CTRL_DOWN_MASK | ALT_DOWN_MASK));
+    keyStrokes.put(NAVIGATE_UP, keyStrokeValue(VK_UP, CTRL_DOWN_MASK | ALT_DOWN_MASK));
+    keyStrokes.put(NAVIGATE_DOWN, keyStrokeValue(VK_DOWN, CTRL_DOWN_MASK | ALT_DOWN_MASK));
+    keyStrokes.put(NAVIGATE_RIGHT, keyStrokeValue(VK_RIGHT, CTRL_DOWN_MASK | ALT_DOWN_MASK));
+    keyStrokes.put(NAVIGATE_LEFT, keyStrokeValue(VK_LEFT, CTRL_DOWN_MASK | ALT_DOWN_MASK));
+
+    return keyStrokes;
   }
 
   private final class ShowHiddenEditPanel implements Control.Command {