+++ changelog.md
@@ -18,6 +18,7 @@ Codion Change Log
 - Control.Command and ActionCommand.perform() renamed execute().
 - KeyEvents.builder(KeyStroke) added.
 - KeyEvents.Builder.enable() and disable() now with varargs.
+- KeyboardShortcut added, first attempt at configurable keyboard shortcuts.
 ### is.codion.framework.domain
 - IdentityKeyGenerator now throws exception in case the generated keys result set is empty.
 ### is.codion.framework.model